[Build Tool] Fixed default NinjaTap plugged device for openMSX [Game Menu] Fixed integer parameter

+++ engine/src/game_menu.c
@@ -188,7 +188,7 @@ void Menu_DisplayItem(u8 item)
 				Print_DrawChar(MENU_CHAR_LEFT);
 			else
 				Print_Space();
-			i16* data = (i16*)pCurItem->Action;
+			i8* data = (i8*)pCurItem->Action;
 			Print_DrawInt(*data);
 			if(g_MenuItem == item)
 				Print_DrawChar(MENU_CHAR_RIGHT);
